+++ sismic/model/__init__.py
@@ -1,4 +1,5 @@
-from .elements import ContractMixin, StateMixin, ActionStateMixin, TransitionStateMixin, CompositeStateMixin, HistoryStateMixin
+from .elements import ContractMixin, StateMixin, ActionStateMixin, \
+    TransitionStateMixin, CompositeStateMixin, HistoryStateMixin
 from .elements import BasicState, CompoundState, OrthogonalState, ShallowHistoryState, DeepHistoryState, FinalState
 from .elements import Transition
 from .statechart import Statechart